Fix all things tox
This commit fixes all the issues with the tox jobs. I initially wrote
this project outside of the openstack infrastructure and played kinda
hard and fast with anything from tox. To correct this I need to add
sphinx docs, a pro forma test to (and the preresiquite testr
boilerplate), and a bunch of random pep8 fixes.

Change-Id: I9946fc7db5c030b5860f03b45471f70c1fc4efb1